jar verify: recognize APK v1 signatures; skip redundant digests
At least some JARs (or APKs in this case) seem to have bogus per-file
sections in the signature file, but Android ignores them because there
is a whole-manifest digest present. Emulate this behavior as long as a
whole-manifest file is present.

Also adds an error if no digest types were recognized at all