Do not return EOF error when reading sector
When reading the sector into the internal buffer, do not return an
`io.EOF` error if encountered. Ignore the error and allow the data to be
copied into the remaining buffer space and the remaining bytes value
updated appropriately. This prevents the reader from receiving an
`io.EOF` early and missing the final available bytes.