Account for padding when patching and reading
When patching __LINKEDIT include any padding that was added when
determining the end of the segment. This fixes the size from being
recorded improperly.

When reading the pages ensure any padding is included in the read. If
the padding is not included in the read, the calculated cdhash value
will be incorrect resulting in an invalid signature.